refactor: UserOrganizationServiceTest 단위 테스트로 수정
