+++ backend/src/test/java/feedzupzup/backend/organization/application/UserOrganizationServiceTest.java
@@ -1,19 +1,15 @@
 package feedzupzup.backend.organization.application;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
 import feedzupzup.backend.config.ServiceIntegrationHelper;
-import feedzupzup.backend.global.exception.ResourceException.ResourceNotFoundException;
 import feedzupzup.backend.organization.domain.Organization;
 import feedzupzup.backend.organization.domain.OrganizationRepository;
 import feedzupzup.backend.organization.dto.request.CheeringRequest;
 import feedzupzup.backend.organization.dto.response.CheeringResponse;
 import feedzupzup.backend.organization.dto.response.UserOrganizationResponse;
 import feedzupzup.backend.organization.fixture.OrganizationFixture;
-import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 
@@ -25,43 +21,26 @@ class UserOrganizationServiceTest extends ServiceIntegrationHelper {
     @Autowired
     private OrganizationRepository organizationRepository;
 
-    @Nested
-    @DisplayName("단체 조회 테스트")
-    class GetOrganizationByIdTest {
+    @Test
+    @DisplayName("단체 ID로 단체를 성공적으로 조회한다")
+    void get_organization_by_id_success() {
+        // given
+        final Organization organization = OrganizationFixture.createAllBlackBox();
 
-        @Test
-        @DisplayName("단체 ID로 단체를 성공적으로 조회한다")
-        void get_organization_by_id_success() {
-            // given
-            final Organization organization = OrganizationFixture.createAllBlackBox();
+        final Organization savedOrganization = organizationRepository.save(organization);
 
-            final Organization savedOrganization = organizationRepository.save(organization);
+        // when
+        final UserOrganizationResponse response = userOrganizationService.getOrganizationByUuid(
+                savedOrganization.getUuid());
 
-            // when
-            final UserOrganizationResponse response = userOrganizationService.getOrganizationByUuid(
-                    savedOrganization.getUuid());
-
-            // then
-            assertThat(response.organizationName()).isEqualTo(organization.getName().getValue());
-            assertThat(response.totalCheeringCount()).isEqualTo(organization.getCheeringCountValue());
-        }
-
-        @Test
-        @DisplayName("존재하지 않는 단체 ID로 조회 시 예외를 발생시킨다")
-        void get_organization_by_id_not_found() {
-            // given
-            final UUID nonExistentOrganizationId = UUID.randomUUID();
-
-            // when & then
-            assertThatThrownBy(() -> userOrganizationService.getOrganizationByUuid(nonExistentOrganizationId))
-                    .isInstanceOf(ResourceNotFoundException.class)
-                    .hasMessage("해당 ID(id = " + nonExistentOrganizationId + ")인 단체를 찾을 수 없습니다.");
-        }
+        // then
+        assertThat(response.organizationName()).isEqualTo(organization.getName().getValue());
+        assertThat(response.totalCheeringCount()).isEqualTo(organization.getCheeringCountValue());
     }
 
     @Test
     @DisplayName("요청한 응원 수만큼 단체의 응원 총 횟수가 증가한다.")
-    void get_organization_by_id_success() {
+    void cheer_increases_total_count() {
         // given
         final Organization organization = OrganizationFixture.createAllBlackBox();
